+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Added
+
+- The `locale`, `mediaFilter`, and `safesearch` props.
+
 ### Changed
 
 - Switch to TypeScript for development.